Add black queen & convert the way moves are defined
Summary:
The moves were defined as a computed earlier but I realised that it's
going to be tricky to be able to determine all the valid moves when the
moves span across the board. So I've defined this as a function instead,
where the board positions are passed & moves for each piece can now be
defined as a math equations instead.